Extract to common component, fix bugs.
