+++ src/pages/DataLibrary.tsx
@@ -1,25 +1,7 @@
 import React, { useEffect, useMemo, useState, useCallback } from 'react'
 import { useNavigate, useParams } from 'react-router-dom'
-import { Box, Skeleton, Typography } from '@mui/material'
-import LibraryTabs from 'src/components/data_library/LibraryTabs'
-import SearchBar from 'src/components/SearchBar'
-import TableHeaderSection from 'src/components/TableHeaderSection'
-import { useLibraryUrlState } from 'src/hooks/useLibraryUrlState'
-import {
-  AssetType,
-  AvailableFilters,
-  ExportableDatasets,
-  LibraryVersionNew,
-  SortOrder,
-  TabConfig,
-} from 'src/types/library'
-import { BioSpecimenType, DatasetTerm, PostMortemIntervalUnit } from 'src/types/model'
-import { assetRegistry } from 'src/components/data_library/assets'
-import LibraryFilters from 'src/components/data_library/LibraryFilters'
-import { useLibraryData, useLibraryMetadata } from 'src/hooks/useLibraryData'
-import LibraryDataGrid from 'src/components/data_library/LibraryDataGrid'
-import { AggregationResult } from 'src/types/elastic'
-import LibraryFooter from 'src/components/data_library/LibraryFooter'
+import { AssetType, ExportableDatasets, LibraryVersionNew, ALL_LIBRARY_TABS } from 'src/types/library'
+import { DatasetTerm } from 'src/types/model'
 import { applyForAccess } from 'src/utils/accessUtils'
 import { getBrandedLibrary } from 'src/libs/libraryVersions'
 import { Storage } from 'src/libs/storage'
@@ -30,79 +12,33 @@ import { TerraDataRepo } from 'src/libs/ajax/TerraDataRepo'
 import { chain, intersection } from 'src/utils/NodashUtil'
 import { EnumerateSnapshotModel, SnapshotSummaryModel } from 'src/types/tdrModel'
 import { getRadarEnabledDatasetsWithRules } from 'src/utils/DatasetUtils'
-import {
-  EMPTY_FILTERS,
-  getFilterSectionsForAsset,
-  sanitizeFiltersForAsset,
-} from 'src/components/data_library/filterRegistry'
-import {
-  clinicalTrialInterventionSelectOptions,
-  clinicalTrialPhaseSelectOptions,
-  clinicalTrialStatusSelectOptions,
-} from 'src/utils/ClinicalTrialEnumUtils'
+import { useLibraryPageState } from 'src/hooks/useLibraryPageState'
+import LibraryPageShell from 'src/components/data_library/LibraryPageShell'
+import LibraryFooter from 'src/components/data_library/LibraryFooter'
+import TableHeaderSection from 'src/components/TableHeaderSection'
+import SearchBar from 'src/components/SearchBar'
 
-/**
- * DataLibrary Page Component
- *
- * Main page for browsing and searching datasets and studies in DUOS.
- * Features:
- * - Tabbed interface for Studies vs Datasets views
- * - Advanced filtering sidebar
- * - Server-side pagination and sorting
- * - Multi-select with Apply for Access functionality
- * - URL-based state for shareability
- *
- * State Management:
- * - URL state: filters, pagination, sort, search, active tab (managed by useLibraryUrlState)
- * - Server state: data fetching, caching (managed by React Query via useLibraryData)
- * - Metadata fetching for filter options (useLibraryMetadata)
- * - Local UI state: selection tracking (useState)
- * - Exportable datasets state for enabling export functionality (useState)
- * - Radar-enabled datasets state for showing Radar integration (useState)
- */
 export const DataLibrary: React.FC = () => {
   const { query } = useParams()
   const navigate = useNavigate()
-
-  const [urlState, updateUrlState] = useLibraryUrlState()
-
-  const [selectedDatasetIds, setSelectedDatasetIds] = useState<number[]>([])
-  const [exportableDatasets, setExportableDatasets] = useState<ExportableDatasets>({})
-  const [radarEnabledDatasetIds, setRadarEnabledDatasetIds] = useState<Set<number>>(new Set())
 
   const user = Storage.getCurrentUser()
   const institutionId = user?.institution?.id
   const institutionName = user?.institution?.name
 
-  const tabs: TabConfig[] = [
-    { key: AssetType.STUDIES, label: 'Studies' },
-    { key: AssetType.DATASETS, label: 'Datasets' },
-    { key: AssetType.MODELS, label: 'AI Models' },
-    { key: AssetType.WORKSPACES, label: 'Workspaces' },
-    { key: AssetType.CLINICAL_TRIALS, label: 'Clinical Trials' },
-    { key: AssetType.BIOSPECIMENS, label: 'Biospecimens' },
-    { key: AssetType.PUBLICATIONS, label: 'Publications' },
-    { key: AssetType.PRESENTATIONS, label: 'Presentations' },
-    { key: AssetType.INTELLECTUAL_PROPERTY, label: 'Intellectual Property' },
-    { key: AssetType.FUNDING_RESOURCES, label: 'Funding Resources' },
-  ]
-
   useEffect(() => {
-    const init = () => {
-      const key = query === undefined ? '/datalibrary' : query.toLowerCase()
-      if (key === 'myinstitution' && !institutionId) {
-        Notifications.showError({ text: 'You must set an institution in your profile to view the `myinstitution` data library' })
-        navigate('/profile')
-      }
-      if (key === '/datalibrary') {
-        Metrics.captureEvent(eventList.dataLibrary)
-      }
-      else {
-        const brand = key.replaceAll('/', '').toLowerCase()
-        Metrics.captureEvent(eventList.dataLibrary, { brand })
-      }
+    const key = query === undefined ? '/datalibrary' : query.toLowerCase()
+    if (key === 'myinstitution' && !institutionId) {
+      Notifications.showError({ text: 'You must set an institution in your profile to view the `myinstitution` data library' })
+      navigate('/profile')
+      return
     }
-    init()
+    if (key === '/datalibrary') {
+      Metrics.captureEvent(eventList.dataLibrary)
+    }
+    else {
+      Metrics.captureEvent(eventList.dataLibrary, { brand: key.replaceAll('/', '').toLowerCase() })
+    }
   }, [query, institutionId, navigate])
 
   const libraryConfig: LibraryVersionNew = useMemo(() => {
@@ -130,160 +66,17 @@ export const DataLibrary: React.FC = () => {
     }
   }, [query, institutionId, institutionName])
 
-  const { data: metadata, isLoading: isMetadataLoading } = useLibraryMetadata(libraryConfig)
+  const pageState = useLibraryPageState(libraryConfig)
+  const { urlState, data, currentAsset, handleSearchChange } = pageState
 
-  const { data, isFetching, error } = useLibraryData(
-    libraryConfig,
-    urlState.tab,
-    urlState.filters,
-    urlState.query ?? '',
-    { page: urlState.page, pageSize: urlState.pageSize },
-    urlState.sortField && urlState.sortOrder
-      ? { field: urlState.sortField, order: urlState.sortOrder }
-      : undefined,
-  )
-
-  const availableFilters: AvailableFilters = useMemo(() => {
-    const dacAgg = (metadata?.dac as AggregationResult)?.buckets || []
-    const dataTypeAgg = (metadata?.data_type as AggregationResult)?.buckets || []
-
-    const uniqueValues = (values: Array<string | undefined | null>) =>
-      [...new Set(values.map(value => value?.trim()).filter(Boolean) as string[])]
-        .sort((a, b) => a.localeCompare(b))
-        .map(value => ({ value, label: value }))
-
-    const workspaceItems = urlState.tab === AssetType.WORKSPACES ? data?.items as Array<{ tools?: string[], platform?: string }> : []
-    const clinicalTrialItems = urlState.tab === AssetType.CLINICAL_TRIALS
-      ? data?.items as Array<{ registry?: string }>
-      : []
-    const biospecimenItems = urlState.tab === AssetType.BIOSPECIMENS
-      ? data?.items as Array<{ optionalDataUse?: string }>
-      : []
-
-    const workspaceTools = uniqueValues(workspaceItems?.flatMap(item => item.tools || []) || [])
-    const workspacePlatform = uniqueValues(workspaceItems?.map(item => item.platform) || [])
-    const clinicalTrialRegistry = uniqueValues(clinicalTrialItems?.map(item => item.registry) || [])
-    const biospecimenDataUse = uniqueValues(biospecimenItems?.map(item => item.optionalDataUse) || [])
-
-    return {
-      accessManagement: [
-        { value: 'open', label: 'Open Access' },
-        { value: 'controlled', label: 'via DUOS' },
-        { value: 'external', label: 'External to DUOS' },
-      ],
-      dataUse: [
-        { value: 'HMB', label: 'Health/Medical/Biomedical' },
-        { value: 'GRU', label: 'General Research Use' },
-        { value: 'DS', label: 'Disease Specific' },
-        { value: 'OTHER', label: 'Other Restriction' },
-        { value: 'NRES', label: 'No Restrictions' },
-      ],
-      dataType: dataTypeAgg
-        .map(bucket => ({
-          value: bucket.key as string,
-          label: bucket.key as string,
-          count: bucket.doc_count,
-        }))
-        .sort((a, b) => a.label.localeCompare(b.label)),
-      dac: dacAgg
-        .map(bucket => ({
-          value: bucket.key as string,
-          label: bucket.key as string,
-          count: bucket.doc_count,
-        }))
-        .sort((a, b) => a.label.localeCompare(b.label)),
-      workspaceTools,
-      workspacePlatform,
-      clinicalTrialStatus: clinicalTrialStatusSelectOptions.map(option => ({ value: option.key, label: option.displayText })),
-      clinicalTrialPhase: clinicalTrialPhaseSelectOptions.map(option => ({ value: option.key, label: option.displayText })),
-      clinicalTrialInterventionType: clinicalTrialInterventionSelectOptions.map(option => ({ value: option.key, label: option.displayText })),
-      clinicalTrialRegistry,
-      biospecimenType: Object.values(BioSpecimenType).map(value => ({ value, label: value })),
-      biospecimenDataUse,
-      biospecimenPostMortemIntervalUnit: Object.values(PostMortemIntervalUnit).map(value => ({ value, label: value })),
-      datasetsCited: [
-        { value: 'true', label: 'Yes' },
-        { value: 'false', label: 'No' },
-      ],
-      biospecimenPostMortemIntervalRange: {
-        min: 0,
-        max: 1000000,
-      },
-      participantCountRange: {
-        min: 0,
-        max: 100000,
-      },
-    }
-  }, [metadata, data?.items, urlState.tab])
-
-  const currentAsset = useMemo(() => assetRegistry[urlState.tab], [urlState.tab])
-  const sanitizedFilters = useMemo(
-    () => sanitizeFiltersForAsset(urlState.tab, urlState.filters),
-    [urlState.tab, urlState.filters],
-  )
-  const filterSections = useMemo(
-    () => getFilterSectionsForAsset(urlState.tab, availableFilters),
-    [urlState.tab, availableFilters],
-  )
+  const [selectedDatasetIds, setSelectedDatasetIds] = useState<number[]>([])
+  const [exportableDatasets, setExportableDatasets] = useState<ExportableDatasets>({})
+  const [radarEnabledDatasetIds, setRadarEnabledDatasetIds] = useState<Set<number>>(new Set())
 
   const selectedStudyIds = useMemo(() => {
     if (!data?.items) return []
-    return currentAsset.getStudyIdsForSelection(
-      data.items,
-      selectedDatasetIds,
-    )
+    return currentAsset.getStudyIdsForSelection(data.items, selectedDatasetIds)
   }, [data, selectedDatasetIds, currentAsset])
-
-  const sortModel = useMemo(() => {
-    if (urlState.sortField && urlState.sortOrder) {
-      return [{ field: urlState.sortField, sort: urlState.sortOrder }]
-    }
-    return []
-  }, [urlState.sortField, urlState.sortOrder])
-
-  const handleTabChange = useCallback((newAssetType: AssetType) => {
-    updateUrlState({
-      tab: newAssetType,
-      page: 0,
-      filters: sanitizeFiltersForAsset(newAssetType, urlState.filters),
-    })
-    setSelectedDatasetIds([])
-  }, [updateUrlState, urlState.filters])
-
-  const handleSearchChange = useCallback((query: string) => {
-    updateUrlState({
-      query,
-      page: 0,
-    })
-  }, [updateUrlState])
-
-  const handleFiltersChange = useCallback((newFilters: typeof urlState.filters) => {
-    updateUrlState({
-      filters: sanitizeFiltersForAsset(urlState.tab, newFilters),
-    })
-  }, [updateUrlState, urlState])
-
-  const handleClearFilters = useCallback(() => {
-    updateUrlState({
-      filters: sanitizeFiltersForAsset(urlState.tab, EMPTY_FILTERS),
-      page: 0,
-    })
-  }, [updateUrlState, urlState.tab])
-
-  const handleSortChange = useCallback((model: Array<{ field: string, sort: SortOrder | null }>) => {
-    if (model.length > 0 && model[0].sort) {
-      updateUrlState({
-        sortField: model[0].field,
-        sortOrder: model[0].sort,
-      })
-    }
-    else {
-      updateUrlState({
-        sortField: undefined,
-        sortOrder: undefined,
-      })
-    }
-  }, [updateUrlState])
 
   const handleSelectionChange = useCallback((datasetIds: number[]) => {
     setSelectedDatasetIds(datasetIds)
@@ -292,10 +85,6 @@ export const DataLibrary: React.FC = () => {
   const handleApplyForAccess = () => {
     applyForAccess(selectedDatasetIds, navigate)
   }
-
-  const handleToggleFilters = useCallback(() => {
-    updateUrlState({ hideFilters: !urlState.hideFilters })
-  }, [updateUrlState, urlState.hideFilters])
 
   useEffect(() => {
     const fetchExportable = async () => {
@@ -349,123 +138,39 @@ export const DataLibrary: React.FC = () => {
     fetchRadarEnabled()
   }, [data?.items, urlState.tab])
 
-  if (error) {
-    return (
-      <Box sx={{ px: 3, py: 4 }}>
-        <Box sx={{ textAlign: 'center', color: 'error.main' }}>
-          <h2>Error Loading Data</h2>
-          <p>{error instanceof Error ? error.message : 'An unexpected error occurred'}</p>
-        </Box>
-      </Box>
-    )
-  }
+  const header = (
+    <>
+      <TableHeaderSection
+        icon={libraryConfig.icon ? { src: libraryConfig.icon } : undefined}
+        title={libraryConfig.title}
+        description={libraryConfig.description}
+      />
+      <SearchBar
+        handleSearchChange={handleSearchChange}
+        initialValue={urlState.query ?? ''}
+        style={{ paddingTop: '10px' }}
+      />
+    </>
+  )
 
   return (
-    <Box sx={{ display: 'flex', flexDirection: 'column', height: '100vh', pb: 5 }}>
-      {/* Header */}
-      <Box>
-        <TableHeaderSection
-          icon={libraryConfig.icon ? { src: libraryConfig.icon } : undefined}
-          title={libraryConfig.title}
-          description={libraryConfig.description}
+    <LibraryPageShell
+      pageState={pageState}
+      tabs={ALL_LIBRARY_TABS}
+      header={header}
+      gridExtras={{
+        selectedDatasetIds,
+        onSelectionChange: handleSelectionChange,
+        exportableDatasets,
+        radarEnabledDatasetIds,
+      }}
+      footer={(
+        <LibraryFooter
+          selectedDatasetIds={selectedDatasetIds}
+          selectedStudyIds={selectedStudyIds}
+          onApplyForAccess={handleApplyForAccess}
         />
-        <SearchBar
-          handleSearchChange={handleSearchChange}
-          initialValue={urlState.query ?? ''}
-          style={{
-            paddingTop: '10px',
-          }}
-        />
-      </Box>
-
-      {/* Tabs */}
-      <Box sx={{ px: 3, pt: 2 }}>
-        <LibraryTabs
-          value={urlState.tab}
-          onChange={handleTabChange}
-          tabs={tabs}
-        />
-      </Box>
-
-      {/* Main content area with filters and grid */}
-      <Box sx={{ display: 'flex', flex: 1, overflow: 'hidden', px: 3, pt: 2 }}>
-        {/* Filters Sidebar */}
-        <Box
-          sx={{
-            width: urlState.hideFilters ? 40 : 280,
-            flexShrink: 0,
-            pr: urlState.hideFilters ? 0 : 2,
-            overflowY: urlState.hideFilters ? 'hidden' : 'auto',
-            overflowX: 'hidden',
-            transition: 'width 0.2s ease',
-          }}
-        >
-          <LibraryFilters
-            filters={sanitizedFilters}
-            onChange={handleFiltersChange}
-            onClear={handleClearFilters}
-            sections={filterSections}
-            loading={isMetadataLoading}
-            isOpen={!urlState.hideFilters}
-            onToggle={handleToggleFilters}
-          />
-        </Box>
-
-        <Box sx={{ flex: 1, height: '100%', overflow: 'hidden', display: 'flex', flexDirection: 'column' }}>
-          {/* Asset count */}
-          {isFetching
-            ? <Skeleton variant="text" width={120} sx={{ fontSize: '15px', mb: 1 }} />
-            : (
-                <Typography
-                  sx={{
-                    color: '#00609f',
-                    fontFamily: 'Montserrat, sans-serif',
-                    fontSize: '15px',
-                    fontWeight: 'bold',
-                    mb: 1,
-                  }}
-                >
-                  {(data?.total ?? 0).toLocaleString()}
-                  {' '}
-                  {data?.total === 1
-                    ? currentAsset.label.singular
-                    : currentAsset.label.plural}
-                </Typography>
-              )}
-          {/* Data Library */}
-          <Box sx={{ flex: 1, minHeight: 0, overflow: 'hidden' }}>
-            <LibraryDataGrid
-              assetType={urlState.tab}
-              data={data?.items || []}
-              loading={isFetching}
-              total={data?.total || 0}
-              paginationModel={{
-                page: urlState.page,
-                pageSize: urlState.pageSize,
-              }}
-              onPaginationChange={(model) => {
-                updateUrlState({
-                  page: model.page,
-                  pageSize: model.pageSize,
-                })
-              }}
-              sortModel={sortModel}
-              onSortChange={handleSortChange}
-              selectedDatasetIds={selectedDatasetIds}
-              onSelectionChange={handleSelectionChange}
-              exportableDatasets={exportableDatasets}
-              radarEnabledDatasetIds={radarEnabledDatasetIds}
-            />
-          </Box>
-        </Box>
-      </Box>
-
-      {/* Footer (shown when assets are selected) */}
-      <LibraryFooter
-        selectedDatasetIds={selectedDatasetIds}
-        selectedStudyIds={selectedStudyIds}
-        onApplyForAccess={handleApplyForAccess}
-      />
-    </Box>
+      )}
+    />
   )
 }
